Deal with tar/compressed files
Currently we are only dealing with directories but we may need to deal
with tar files and those tar files can also be compressed with gzip or
zstd.

This patch makes it so we can deal withthose files both on generate and
check by uncompressing them onto a temp directory and
generating/checking against those files.

It also fixes the keywords used by removing the time and size from the
default keywords and appends any given keywords to the check

Signed-off-by: Itxaka <igarcia@suse.com>