Add event dispatcher
Currently we want to trigger actions on luet events, so we can act as a
plugin for luet without having to manually trigger actions.

This patch adds a root command that triggers when called with no
subcommand which parses an event and a payload and dispatches it to the
correct path.

Currently only the unpack event is supported, which triggers a download
of the metadata, extraction of the mtree checksums and triggers a check
of the validation files against the actual files

Signed-off-by: Itxaka <igarcia@suse.com>