Fix using the same instance of values during the tests
I was using just one instance of DefaultReturnValue to store the
results, which seemed to be ok when launching it via CLI as the program
terminated and the values were cleared.

But on testing, because we are calling the unpack several times the same
instance was around and was getting polluted via other tests, getting
strange errors when the state was a success.

This patch reworks it into being unique each time and also generating it
on demand from the WrapErrorMap helper, so we have to write less code :)

Signed-off-by: Itxaka <igarcia@suse.com>