Drop xattrs and allow for more hands-on with keywords
Drop teh default xattrs from the generation as this has probed
troublesome when working with docker, as it doesnt keep some file
capabilities so the mtree checks never succedes.

Also adds a new option -K to override the keyword list. This will allow
us to control the keywords used in the future as it allows us not only
to expand the default keywords (sha512digest and type) but also to fully
override the list of keywords via command line instead of having to
modify the code to add those.

The default ones left in the code are the basic ones that we need, if
something needs to drop those, then the mtree check is utterly useless.

Signed-off-by: Itxaka <igarcia@suse.com>